Changedthelandingspagesothatitwillworkbetter
The landing page now scrolls instead of going to the next section which will make it available to view all the text on the webpage. The how we do it section now also centers on mobile devices which will make it more readable for the mobile users. Logo of EHVLINC is still not rendable formidable cansable.